updatesinhvien/vubakiem

+++ database/migrations/2025_10_25_182019_add_loai_id_to_kho_tai_san_table.php
@@ -10,19 +10,21 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('kho_tai_san', function (Blueprint $table) {
-        $table->unsignedBigInteger('loai_id')->after('id'); // hoặc sau cột nào phù hợp
-        $table->foreign('loai_id')->references('id')->on('loai_tai_san')->onDelete('cascade');
-    });
-}
+    {
+        Schema::table('kho_tai_san', function (Blueprint $table) {
+            if (!Schema::hasColumn('kho_tai_san', 'loai_id')) {
+                $table->unsignedBigInteger('loai_id')->nullable()->after('id');
+                $table->foreign('loai_id')->references('id')->on('loai_tai_san')->onDelete('cascade');
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('kho_tai_san', function (Blueprint $table) {
-        $table->dropForeign(['loai_id']);
-        $table->dropColumn('loai_id');
-    });
-}
 
+    public function down()
+    {
+        Schema::table('kho_tai_san', function (Blueprint $table) {
+            $table->dropForeign(['loai_id']);
+            $table->dropColumn('loai_id');
+        });
+    }
 };
